--- cuenta_yo/off-chain/wallets/enterprise.vkey
@@ -1,5 +1,0 @@
-{
-    "type": "PaymentVerificationKeyShelley_ed25519",
-    "description": "Payment Verification Key",
-    "cborHex": "582003aec869f0868a0fdfdcb7f45bf36c21d5626f1f6666341e34586ead94dda10a"
-}
