--- cuenta_yo/off-chain/wallets/payment.addr
@@ -1,1 +1,0 @@
-addr_test1qpp0ppfz5y8j22tf57nguhnv2lzqp2r2nuy7plurrtx5srpy8y6r9xqh29l009qdjdfdse0qahggcftxnd8wsd9f4glqxlp6qg
