+++ cuenta_yo/plutus.json
@@ -1,0 +1,64 @@
+{
+  "preamble": {
+    "title": "rcerrud/cuenta_yo",
+    "description": "Aiken contracts for project 'rcerrud/cuenta_yo'",
+    "version": "0.0.0",
+    "plutusVersion": "v2",
+    "compiler": {
+      "name": "Aiken",
+      "version": "v1.0.29-alpha+16fb02e"
+    },
+    "license": "Apache-2.0"
+  },
+  "validators": [
+    {
+      "title": "cuenta_yo.cuenta_yo",
+      "datum": {
+        "title": "datum",
+        "schema": {
+          "$ref": "#/definitions/cuenta_yo~1DatumYO"
+        }
+      },
+      "redeemer": {
+        "title": "redeemer",
+        "schema": {
+          "$ref": "#/definitions/Int"
+        }
+      },
+      "compiledCode": "590165010000323232323232232232253330063253330073370e900118041baa300130093754004264a66601066e1cdd6980098051baa00700413233232232533300d3370e900000089999112999808198038018008a51133007002001375c600a601e6ea8030dd7180398079baa00c37586006601e6ea800854ccc034cdc3a40040022666444a6660206600e00600226600e0040022940dd7180298079baa00c375c600e601e6ea8030dd6180198079baa00214a06eb4c014c038dd500591807980818081808180818081808180818080009119198008008011129998080008a50132533300e3371e6eb8c0480080145288998018018009809000980098059baa0042300d00114a046018601a601a00229408c02cc03000452613656375a002a66600466e1d2000300337540022646464646464a666016601a0042930b1bad300b001300b002375c601200260120046eb8c01c004c010dd50008b2b9a5573aaae7955cfaba15745",
+      "hash": "3777c18154eeae505ba4bd4da1c817530d570da92cdac2e1ca8a1516"
+    }
+  ],
+  "definitions": {
+    "ByteArray": {
+      "dataType": "bytes"
+    },
+    "Int": {
+      "dataType": "integer"
+    },
+    "cuenta_yo/DatumYO": {
+      "title": "DatumYO",
+      "anyOf": [
+        {
+          "title": "DatumYO",
+          "dataType": "constructor",
+          "index": 0,
+          "fields": [
+            {
+              "title": "firmante_1",
+              "$ref": "#/definitions/ByteArray"
+            },
+            {
+              "title": "firmante_2",
+              "$ref": "#/definitions/ByteArray"
+            },
+            {
+              "title": "tipo_cuenta",
+              "$ref": "#/definitions/Int"
+            }
+          ]
+        }
+      ]
+    }
+  }
+}
