+++ nativetokens/aiken.lock
@@ -3,12 +3,12 @@
 
 [[requirements]]
 name = "aiken-lang/stdlib"
-version = "1.8.0"
+version = "1.9.0"
 source = "github"
 
 [[packages]]
 name = "aiken-lang/stdlib"
-version = "1.8.0"
+version = "1.9.0"
 requirements = []
 source = "github"
