+++ nativetokens/plutus.json
@@ -6,7 +6,7 @@
     "plutusVersion": "v2",
     "compiler": {
       "name": "Aiken",
-      "version": "v1.0.26-alpha+075668b"
+      "version": "v1.0.29-alpha+16fb02e"
     },
     "license": "Apache-2.0"
   },
@@ -32,7 +32,7 @@
       },
       "parameters": [
         {
-          "title": "txOutRef",
+          "title": "uniqueUTxO",
           "schema": {
             "$ref": "#/definitions/aiken~1transaction~1OutputReference"
           }
@@ -44,8 +44,8 @@
           }
         }
       ],
-      "compiledCode": "5902ab0100003232323232323222322323232253330093232533300b3006300c375400226464a66601a64660020026eb0c04cc040dd5180118081baa00622533301200114a0264a66602066ebcc010c048dd5180a8010078a5113300300300130150011533300d3232533301230150021325333010300a375a6024004266e3c0040345281bae301000116301300132533300e3008300f3754002297adef6c6013756602660206ea8004c8cc004004c8cc004004dd5980a180a980a980a980a98089baa30033011375400e44a666026002297adef6c6013232323253330143371e911000021003133018337606ea4008dd3000998030030019bab3015003375c6026004602e004602a00244a666024002298103d87a8000132323253330123371e00e6eb8c04c00c4cdd2a40006602c6e980052f5c026600a00a0046eacc04c008c058008c05000454ccc034c020c038dd5180098079baa0091332232330010013758602a602c602c602c602c602c602c602c602c60246ea800c894ccc050004528099299980919b8f375c602e00400829444cc00c00c004c05c004c004c03cdd50029bae301230133013300f37540122664464a666020601460226ea80044cdc48011bad3015301237540022940c00cc044dd5180318089baa00230123013301330133013301330133013300f37546002601e6ea8014dd6980218079baa00914a029408c048004dd7180818069baa001163001300c37540044601e602000229309b2b2999803980118041baa0031323232323232323253330123015002132498c94ccc040c02c00454ccc04cc048dd50040a4c2c2a66602060140022a66602660246ea802052616163010375400e2c6eb4c04c004c04c008dd7180880098088011bad300f001300f002300d001300937540062c6e1d2002370e90001bae0015734aae7555cf2ab9f5740ae855d11",
-      "hash": "8f3ab74cb362109b534b56e03abad2abe908f30cee17cca917267926"
+      "compiledCode": "5902a801000032323232323232223223232253330083232533300a3005300b375400226464a66601864660020026eb0c048c03cdd5180118079baa00622533301100114a0264a66601e66ebcc010c044dd5180a0010070a511330030030013014001153323300d32325333012301500213253330103004375a6024004266e3c0040345281bae301000116301300132533300e3002300f3754002297adef6c6013756602660206ea8004c8cc004004c8cc004004dd5980a180a980a980a980a98089baa30043011375401044a666026002297adef6c6013232323253330143372291100002153330143371e91010000210031005133018337606ea4008dd3000998030030019bab3015003375c6026004602e004602a00244a666024002298103d87a80001323232325333013337220120042a66602666e3c0240084cdd2a40006602e6e980052f5c02980103d87a8000133006006003375660280066eb8c048008c058008c0500044c94ccc038c0240044cc88c8cc004004dd6180b180b980b980b980b980b980b980b980b98099baa00322533301500114a0264a66602666e3cdd7180c0010020a51133003003001301800130033010375400e6eb8c04cc050c050c040dd50050a999807180100089991192999808980298091baa0011337120046eb4c058c04cdd50008a50300530123754601060246ea8008c04cc050c050c050c050c050c050c050c040dd5180198081baa007375a600c60206ea80285281bad3002300f37540126e1d200214a029408c044004dd7180798061baa001163001300b37540044601c601e00229309b2b2999803180098039baa0021323232323232323253330113014002149858dd6980900098090011bae30100013010002375a601c002601c0046eb4c030004c020dd50010b1b8748000dd7000ab9a5573aaae7955cfaba05742ae89",
+      "hash": "090b19a9a7c400dd14a0dae42f9cf21e165d9a1dd1f815078c6f8a0a"
     }
   ],
   "definitions": {
@@ -138,23 +138,6 @@
         }
       ]
     },
-    "eanft/Action": {
-      "title": "Action",
-      "anyOf": [
-        {
-          "title": "Owner",
-          "dataType": "constructor",
-          "index": 0,
-          "fields": []
-        },
-        {
-          "title": "Time",
-          "dataType": "constructor",
-          "index": 1,
-          "fields": []
-        }
-      ]
-    },
     "eanft/ConditionsRedeemer": {
       "title": "ConditionsRedeemer",
       "anyOf": [
@@ -165,7 +148,7 @@
           "fields": [
             {
               "title": "action",
-              "$ref": "#/definitions/eanft~1Action"
+              "$ref": "#/definitions/Int"
             },
             {
               "title": "deadline",
